improve CurvilinearGrid

+++ include/vapor/utils.h
@@ -106,12 +106,12 @@ COMMON_API size_t VProduct(const std::vector<size_t> &a);
 
 //! Vectorize a coordinate offset. Inverse of VectorizeLinearize
 //!
-COMMON_API void VectorizeCoords(size_t offset, const size_t *min, const size_t *max, size_t *coords, int n);
+COMMON_API void VectorizeCoords(size_t offset, const size_t *min, const size_t *max, size_t *coords, size_t n);
 COMMON_API std::vector<size_t> VectorizeCoords(size_t offset, const std::vector<size_t> &min, const std::vector<size_t> &max);
 
 //! Vectorize a coordinate offset. Inverse of VectorizeLinearize
 //!
-COMMON_API void VectorizeCoords(size_t offset, const size_t *dims, size_t *coords, int n);
+COMMON_API void VectorizeCoords(size_t offset, const size_t *dims, size_t *coords, size_t n);
 COMMON_API std::vector<size_t> VectorizeCoords(size_t offset, const std::vector<size_t> &dims);
 
 //
